refactor(sse): extrai quota-core (UsageQuota + builders) de services/usage.ts
Continua a decomposição do god-file services/usage.ts (empilhado sobre #4949).
Move o quota-core compartilhado por todos os fetchers de uso para
services/usage/quota.ts:

- type UsageQuota — a forma de valor que todo fetcher por-provider retorna.
- parseResetTime — normaliza timestamps de reset upstream (segundos vs millis,
  epoch-zero → null, strings numéricas). Puro.
- createQuotaFromUsage — monta uma UsageQuota a partir de used/total (clamp +
  remaining %), importando os escalares do leaf irmão (./scalars).

Pré-requisito para extrair as famílias de provider (MiniMax/GLM/Antigravity) como
leaves: agora cada família pode importar UsageQuota/parseResetTime/createQuotaFromUsage
do leaf neutro em vez de puxá-los do host (que importaria o fetcher de volta → ciclo).

Behavior-preserving (byte-idêntico). usage.ts 3222 → 3147.

Validação: typecheck:core, eslint 0/0, check:cycles, check:file-size,
usage-service-hardening (24/24) + caracterização nova (3/3, normalização de
timestamp + assembly de quota).
